Actualizo a versión 5.0 (refactor completo)

+++ SmartPassBack-master/SmartPassBack-main/smartpass-backend/src/main/java/com/smartpass/smartpassbackend/controller/TransitoController.java
@@ -23,7 +23,7 @@ public class TransitoController {
     }
 
     @GetMapping("/cliente/{idCliente}")
-    public List<Transito> obtenerPorCliente(@PathVariable int idCliente) {
+    public List<Transito> obtenerPorCliente(@PathVariable Long idCliente) {
         return service.obtenerPorCliente(idCliente);
     }
 }
