+++ SmartPassBack-master/SmartPassBack-main/smartpass-backend/src/main/java/com/smartpass/smartpassbackend/service/TransitoService.java
@@ -19,7 +19,9 @@ public class TransitoService {
         return repository.findAll();
     }
 
-    public List<Transito> obtenerPorCliente(int idCliente) {
+    public List<Transito> obtenerPorCliente(Long idCliente) {
         return repository.findByIdCliente(idCliente);
     }
+
+
 }
